修改：map.html 大地圖新增

+++ FoodCar/src/main/java/model/MenusVO.java
@@ -8,7 +8,7 @@ public class MenusVO implements java.io.Serializable {
 	private Integer sID ;
 	private Integer kindID ;
 	private Integer statusID ;
-	private Double unitPrice ;
+	private Integer unitPrice ;
 	private String foodIMG ;
 	private StoresVO stroeVO;
 	
@@ -48,10 +48,10 @@ public class MenusVO implements java.io.Serializable {
 	public void setStatusID(Integer statusID) {
 		this.statusID = statusID;
 	}
-	public Double getUnitPrice() {
+	public Integer getUnitPrice() {
 		return unitPrice;
 	}
-	public void setUnitPrice(Double unitPrice) {
+	public void setUnitPrice(Integer unitPrice) {
 		this.unitPrice = unitPrice;
 	}
 	public String getFoodIMG() {
